First pass at IamRoleElevatedManagedPolicyRule.rb
