+++ spec/test_templates/json/iam_role/iam_role_with_poweruseraccess_policy.json
@@ -1,0 +1,22 @@
+{
+    "AWSTemplateFormatVersion": "2010-09-09",
+    "Resources": {
+      "RoleWithPowerUserAccessPolicy": {
+        "Type": "AWS::IAM::Role",
+        "Properties": {
+          "AssumeRolePolicyDocument": {
+            "Version" : "2012-10-17",
+            "Statement": [ {
+              "Effect": "Allow",
+              "Principal": {
+                "Service": [ "ec2.amazonaws.com" ],
+                "AWS" : "arn:aws:iam::324320755747:root"
+              },
+              "Action": ["sts:AssumeRole"]
+            } ]
+          },
+          "ManagedPolicyArns" : [ "arn:aws:iam::aws:policy/PowerUserAccess" ]
+        }
+      }
+    }
+  }
